+++ types/roll-bonus-api.d.ts
@@ -207,7 +207,7 @@ import { signed } from '../src/util/to-signed-string.mjs';
 import { Trait } from '../src/util/trait-builder.mjs';
 import { truthiness } from '../src/util/truthiness.mjs';
 import { distinct, uniqueArray } from '../src/util/unique-array.mjs';
-import { WhileWeaponTypeTarget } from '../src/targeted/targets/conditional/while-weapon-type-equipped-target.mjs';
+import { WhileWeaponTypeTarget } from '../src/targeted/targets/conditional/while-equipped-target.mjs';
 
 export class _RollBonusesAPI {
     es: any;