+++ changelog.md
@@ -13,8 +13,8 @@
   - Target tokens based off of their size relative to your own current size.
 - **Size Target**
   - Activate bonus when targeting a token of a specific size.
-- **While Weapon Type Equipped**
-  - Activates bonus while a specified weapon type is equipped (this is useful for cases like "+2 Intimidate when using a Longsword)
+- **While Equipped**
+  - Activates bonus while a specified weapon or armor type is equipped (or while having a specified natural attack type) (this is useful for cases like "+2 Intimidate when using a Longsword)
     - Do not confuse this with "Weapon Type Target" which grants bonuses only to specific weapons -- this will grant bonuses _even if using a different weapon_ as long as you have the specified weapon type equipped in your inventory tab.
 
 ### Improvements